file: default constructor
Some code is written to assume that a "file" member can be default-
constructed and only set later when we manage to open a file.
This is especially true for code which was previously written with
lw_shared_ptr<file>.

So this patch adds a default constructor for "file", and an operator bool
which can be used to check if the file is set or not.

Actually *using* a default-constructed file can cause a null pointer
dereference - just like using a default-constructed lw_shared_ptr.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>