build: strip all tests
This patch saves about 1.5 GB of disk space in Seastar's build directory,
as well as a lot of memory during the link phase of the build (which can
be noticable on low-memory machines which leads to slow swapping).

Because of C++'s extremely lengthy mangled names, and extremely numerous
functions, the debugging information generated for Seastar code is absurdly
large, and added to every single executable we generate. This is most
noticable in tests - we currently have about 30 tests (with hopefully much
more to come), each compiled into a separate excutable with its own copy of
all this debug information.

So this patch creates all test executables - whether debug or release mode -
stripped. When a user encounters a failing test he wants debug information
for (for gdb or the sanitizer), he can trivially relink it unstripped,
with a command like:

    ninja build/release/tests/allocator_test_g

note the added "_g". This links the already existing object files (which
still have their debug information, which takes just a fraction of a second.

On my machine, this patch reduces the Seastar built tests from about
2.8 GB to less than 1.3 GB. The build/release/tests directory drops from
1 GB to just 0.2 GB! The build/debug/tests directory is still huge (1.8 GB)
but this happens not because of the debug information, but because of the
undefined-behavior sanitizer (-fsanitize=undefined), which unfortunately
adds a huge data segment to each executable and I still don't know how to
improve on that. Nevertheless, it's still a significant reduction in space
and will be even more important as we write more tests.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>