sharded: avoid SIGSEGV on stop
If start() throws an exception, stop will eventually fail with SIGSEGV. In
particular, the instances that we didn't have the chance to initialize will be
zeroed, and when stop() will try to dereference it, this is where the problem
occurs.

Not calling stop is usually not an option: it might be the case that some
instances are initialized properly, and some are not. Also, if we don't even
register a stop function, we will crash with the all-too-common "_instances not
empty" assert.

The caller has absolutely no way to know what to do in this situation, and I
will argue that everything should be handled inside the primitive.

So if we fail, the best way to do is make sure the array is properly destroyed.
This will guarantee that the instances already constructed are destroyed, at
the same time that will present stop with an empty instances array - pretty
much nullifying its loop.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>