sharded: fix map() method
The mapper function that passed to the map function should be catched as
value not as a referece, or it could get out of scope and deleted which
will result in a segmentation fault.

While fixing that, the result vector will be return with std::move to
prevent unneeded copy.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>