dpdk: Linearize too fragmented xmit packet for i40e NICs
Intel's 40G NICs have HW limitation for a maximum number of fragments
in a non-TSO xmit packet and to the number of fragments on which the
TSO segment may span. They may have not more than 8 fragments including
TSO header, which is accounted as a separate fragment.

This limitation should have been handled by the device driver however
DPDK (or the current Linux driver) doesn't handle it at the moment.

This patch introduces a temporary workaround for this issue by linearizing
the xmit packet if the underlying device is 40G Intel NIC and the packet
is as described above.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>