net: prevent tcp from fragmenting packet headers
tcp calls packet::share() to store a data copy in the unacked queue.  It
does so, however, after prepending the tcp header (but taking care not
to copy it).  Unfortunately, packet::share() will call unuse_internal_data()
which will copy the tcp header from the header area to the data area -- on
the original packet.

The original packet is then passed on to the IP layer, which prepends its
own header.  But because of the earlier call to unuse_internal_data(), these
headers will not be contiguous, causing more work for the NIC.

Fix by calling packet::share() earlier, before the tcp header is prepended.
Now the call to unuse_internal_data() does nothing, because there is nothing
to unuse.