net::dpdk: call rte_pktmbuf_reset() when recycling the mbuf
When packet is recycled it won't get to the mempool and thus won't
be allocated from it as the deleted comment claimed - it will be
allocated directly from the tx_buf_factory's cache. Therefore
resetting just the 'next' field is not enough and rte_pktmbuf_reset()
should be called for the mbuf before it's pushed into the cache.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>