net: make udp send more robust wrt. errors
udp uses a semaphore to limit the number of inflight packets per channel,
adding the number of bytes back when the packet is moved to L2.  However,
if an exception in encounterd (arp?) then the sempahore will not be
incremented, and eventually the udp channel will shut down.

Fix by using the packet deleter mechanism to increment the semaphore instead.
This alters the meaning of in-flight, so larger buffers may be needed.