future: finally(): don't let the exceptional future to be ignored
If the original return value was an exceptional future and the
return of the callback is also an exceptional future, then the
original result will be ignored causing the well-known annoying
(by righteous) message to be printed.

Instead, when we are about to return an exceptional future returned
by a callback, consume the original future and dicard any value
it was holding.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>