future: improve handle_exception
This patch makes future<T...>::handle_exception more general and useful,
as suggested by Avi:

Before this patch, future<T...>::handle_exception() returned a future<>.
In other words, i.e., it discarded the future's value, even on a successful.
future.

After this patch, the value is no longer automatically discarded -
future<T...>::handle_exception() returns future<T...>. Instead of
discarding the value, the exception handling callback is required to
return a replacement value (which can be an immediate value, or another
future) to replace the exception.

For example how the improved handle_exception can be used:

	extern future<result> calculate_result();
	result exceptional_result(std::exception_ptr eptr) {
		// create some replacement result based on the
		// exception
		...
	}
	return calculate_result().handle_exception(exceptional_result);

For future<>, there is no change and the exception handling functor
doesn't need to return anything. For a future<T...> with non-empty value,
if the user wants to discard the value as before, he needs to do this
explicitly, with

          fut.discard_result().handle_exception(....);

Note how in this example handle_exception() runs on a future<>, so the
exception-handling function doesn't need to return anything - as in
before this patch.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>