httpd: fix future exception handling
[in v2: use new handle_exception()]

When the request handler throws an exception, the existing code correctly
converts this into the appropriate HTTP failure (error code, message, etc.).
However, the handler can successfully return a future, but that future can
be *later* be resolved with an exception - and we forgot to handle this case.
When this happens, the client gets a completely empty response, instead of
an HTTP error, and an ignored exceptional future is written to the log.

This patch fixes this. Because the exception-processing code now needs
to be done in two places (for an exception thrown by the handler, or
for an exception later set on the future returned by the handler), this
code is moded into a separate function.

This patch also fixes another bug: the code previously only captured
exceptions of subclasses of std::exception, and other exceptions could
cause the entire server to crash. This code also catches exceptions of any
("...") type, and generates a generic message "unknown unhandled exception".

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>