future: add get_value(), and use it in then()
After then() realizes the future has a value (not an exception), it uses
get() to get this value. But get() has extra code to handle an exception,
which we already know is not the case.

So this patch adds a new future_state method - get_value() - to get the
value when we already know the future is available() and !failed(). It is
the counterpart to the get_exception() function we already have.

We then use this new method in then().

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>