future: don't use get() in future_state::forward_to()
Our implementation of forward_to() used get() when it already knows this
future has a value, not an exception, so better just use the value directly
than call get(). It is possible the compiler can optimize away most of
get()'s code in this case, but why give the compiler more work?

Also, with the new code, there is more symmetry between the value and
exception cases.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>