json2code: Add enum support in return type
Sometimes it is usefull to return an enum as a direct type, so far it
was supported when the enum was part of a module (ie. an object). This
change add enum support outside of a module.

The enum support is based on an enum declaration and a wrapper class
that can be safely assigned from other enums allowing a simple type
conversion between internal enums and the API.

The main change is making a function out of the enum wrapper creation.
When moving the code, the multiple method calls where replaced with
string Template

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>