memory: try harder to allocate large blocks
We use index_of_conservative() so that the search for a free block is
guaranteed to run quickly (it ensures that the lists we search have free
blocks at least as large as the requested block), but that can leave
us with unsatisfied allocations even though enough a free block of the
required size is available, but simply in a list of smaller blocks.

Fix by detecting the situation, and falling back to a slower search if we
reach it.