sharded: do not capture remote service pointer on remote invocation lambda
Current sharded<> implementation captures remote service pointer in a
lambda, but service may be stopped before lambda is executed in which
case use-after-free will happen. Fix this by obtaining reference to a
service on a shard that is going to use it and throwing an exception
if it does not exist.