memory: Fix infinite recursion when throwing std::bad_alloc
Fixes #51.

When throwing std::bad_alloc, the runtime is calling
__cxa_allocate_exception, which in turn calls malloc() to allocate the
exception object. This allocation may also fail. libstdc++ has a
fallback emergency pool which it uses in case malloc() returns
nullptr, but if we throw somewhere inside malloc(), we won't get
there, we'll recurse infinitely until SIGSEGV.

The fix is to avoid throwing inside malloc() in case we're out of
memory.