memory: Reclaim memory synchronously with allocation
The goal is to make allocations less likely to fail.

Currently reclaimer runs asynchronously in a separate fiber when the
number of free pages falls below the threshold. We can run out of
memory if the current fiber allocates a lot of memory before it defers
or if it attempts a large allocation which can't be satisfied due to
fragmentation.

This patch changes that by allowing the reclaimer to run synchronously
with allocations. Each reclaimer can decide if it's able to run
synchronously with allocations.

Synchronous reclaimers are run either when the amount of free memory
falls below the threshold or when allocation fails. We always try to
restore the memory above the threshold. When allocation fails, we
reclaim until allocation can be satisfied or reclaimer makes no
forward progress.

The case of satisfying a large allocation is currently handled in a
simplistic way, we simply evict until large enough contiguous block of
memory becomes available. Dependning on how fragmented the memory is,
this may result in evicting a lot of objects. We should therefore
strive to make large allocations as rare as possible, by fragmenting
large allocations.