iostream: support background operations for data_source
To support read-ahead, then we need a way to wait for any read-aheads to
complete after the user is done with the stream.  Add support for this
to data_source.

This data_source_impl::close() API is completely internal; the input_stream
user is not expected to know about read ahead.