future: Fix assertion failure in case schedule() throws
In future::then_wrapped(), when schedule() throws, by the time this
happens `pr` is already moved into the lambda:

    typename futurator::promise_type pr;
    auto fut = pr.get_future();
    try {
        schedule([pr = std::move(pr), func = std::forward<Func>(func)] (auto&& state) mutable {
            futurator::apply(std::forward<Func>(func), future(std::move(state))).forward_to(std::move(pr));
        });
    } catch (...) {
        // catch possible std::bad_alloc in schedule() above
        return futurator::make_exception_future(std::current_exception());
    }
    return fut;

It will die with the lambda before we reach the catch clause,
orphaning `fut`. This triggers an assertion failure inside
promise::abandon(). 'fut' didn't escape the function yet and will die
soon after we return from then_wrapped() via the catch clause, but the
promise doesn't know this.

This change fixes the problem by relaxing restriction checked by
abandon(). Currently it is illegal to destroy an unresolved promise if
there is a live future associated with that promise. It will be
relaxed so that it is illegal to destroy such promise only when there
is a live future associated with it with a scheduled continuation. So
if continuation is never scheduled, we're fine.