Http server: Add an option to register general exception handler
The http has a general exception handler for general behaviour that is
called for all unhandled excecptions.

The general exception handler support three cases, http exception, std
exception and general exception.

Exception that are not http exceptions, are mapped to a 500 reply.

Specific applications that uses seastar, sometimes define their own
specific exceptions, that should be handled in the general exception
handler.

For example an application that throw a specific parameter validation
exception, may want that this exception will be mapped to 400 with an
appropriate message.

This patch allows application that uses the seastar library to register
additional exception handler.

The exception handler function gets an exception ptr and must call
std::rethrow_exception(eptr); and catch the exception it wishes to
handle.

If it does not catch the specific exception, or throw a new exception,
other register exception handler would get the exception.

If after all the handlers there is still an exception either because non
catched the original one, or if a handler catch and throw a new
exception, the basic handling remain: api exceptions are return by their
internal http code and other exceptions are mapped to 500.

This handler does not limit the option of a specific API call to catch
the exceptions itself and add its own logic.

When regisering an exception handler and id is return that can be used
to un-register the handler.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>