thread: scheduling groups
A scheduling groups allows limiting a group of compute-intensive thread's
cpu usage to a certain fraction of the core.  The limit applies over a
specified period, and is reset when the period ends.

For example, a 20% limit with a period of 1ms will limit the thread (or
group of threads) to using 200usec of cpu time in any 1ms period.