memory: fix freeing of very small (<8 byte) objects with sized deallocation
When we allocate, we round up the allocation size to 8 bytes, so we have
room for the free pointer.  The compiler, however, does not match us
when it uses sized deallocation (operator delete(void*, size_t)), so
we have to do it ourselves or we hit the wrong pool.  It doesn't help
that size_to_idx() is broken for small sizes, generating a negative index
(or maybe it did help by exposing this bug).