fix futurize<void> for the case in which Func returns a future
futurize is supposed to turn the result of a call into a function if it is not
already one. For the void case, it fails if the funcion passed is returns a
future.

It will do:
    func(args);
    return make_ready_future<>();

However, if func already returns a future, this means we won't wait for it, and
the code simply breaks.

The typed case deals with this through convert(). But for void it gets harder
because we cannot assign its return value like we do in there.

This patch codes a free function that will use sfinae as a helper to deal with
that.

Fixes #55.