test futurize
futurize keeps breaking, so I went to look for why the test cases for futurize
aren't catching the breaks. Well, it was a surprisingly simple explanation:
they don't catch it, because they don't exist. That's quite shameful for such
core facility.

This test will not catch all possible breakage scenarios, but it will already
test for the most common cases.