stream: workaround native network stack drops
The native network stack expects that stream::produce() return a
ready, unfailed future in order to push the next packet.  However
commit 23f4faeea516426ba ("stream: use a smaller stick for failed
producers") adds a .then_wrapped() call, which sometimes converts
a ready furure to an unready one (due to max_inlined_continuations).

While this needs to be fixed in a better way, for now work around
it by detecting that the future is ready and not failed, and thus
the then_wrapped() call will have no effect, and returning
immediately.