rpc: do not wait for data to be sent
When rpc was streaming it was important to wait until data is send
before making future returned by an rpc call ready. Now there is no
need to wait for data to be send since it is serialized into internal
buffer synchronously. Furthermore waiting for data to be send may delay
timeout reporting. The patch removes waiting for send completion for
each individual message. Instead wait for all messages to be send on
shutdown only.

Tested-by: Asias He <asias@cloudius-systems.com>