reactor: temporarily disable SO_REUSEPORT.
Using SO_REUSEPORT is causing heavy connection imbalance under certain
situations, making some cores overloaded, while others sit free. It is up to
the kernel when to distribute the connections, and that will not necessarily
lead to the ideal case for us.

The proper solution is to employ better balancing mechanisms, some of them
which will probably need the protocol's cooperation. But until we do that,
let's just disable it temporarily.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>