deannoyify touch_directory
Our touch_directory tries to create a directory normally and handles the
exception in case it doesn't exist. It was meant to be used in early setup
cases. Since then, its usage evolved. In projects like ScyllaDB,
touch_directory, and its recursive version, are used in the normal write path,
meaning that frequent operations are now always throwing exceptions left and
right.

Clearly its users have outgrown its original design and it is time for a change:
generating exceptions in common path is at the very least annoying when debugging
and it would be much better if we could avoid it altogether.

This patch does that by creating a specialized touch_directory inside the
reactor itself. It will handle the error case in the thread pool without
creating an exception for this case.

Signed-off-by: Glauber Costa <glommer@scylladb.com>