configure.py: fix use of "echo -e"
The battle between the factions of BSD-inspired "echo -e" and the
SystemV-inspired "echo with escape sequences" has been raging on for
decades, and not even Seastar can avoid getting caught in the fighting ;-)

It turns out that on Ubuntu, the shell defaults to dash, not bash,
and dash has a builtin "echo" command which, unlikes Linux's GNU echo,
does not support the "-e" option. This breaks the generation of the
seastar.pc file.

This patch changes "echo -e" to "/bin/echo -e". This is not a perfect
solution - it assumes that echo is installed in /bin (which seems to be
true on every known Linux system), and that /bin/echo supports "-e" (this
is true on every Linux system with Gnu's utilities, but won't be true on
some antediluvian Unix systems). But it's a good enough patch to fix the
problem on Ubuntu.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>