doc: add Seastar tutorial
This patch puts an updated version of the very partial Seastar tutorial
I once wrote for the wiki, in the doc/ directory inside the Seastar
repository. The tutorial is titled "Asynchronous Programming with Seastar",
and could use a lot more work, but with some love, it has the potential to be
very useful for aspiring authors of Seastar applications.

The document doc/tutorial.md is written in the markdown format, which
is a first for me (usually I'm a TeX aficionado), and so far seems to
be working fairly well: This file can probably be viewed directly with a
markdown viewer (e.g., on Github), but perhaps more interestingly it can be
converted to nicely formatted HTML or PDF with the commands

	ninja doc/tutorial.html
	ninja doc/tutorial.pdf

The "pandoc" command is used for these conversions, so you must have it
installed if you want to do these conversions.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>