fix output stream batching
Current code for output stream batching may execute two writes to
underlying fd in parallel which is illegal. It may happen since both
callback that is called by the poller and output stream user may run in
parallel and do not synchronize with one another. This patch aims to add
such synchronisation. If ostream user writes while poller is running it
will wait for poller to complete, otherwise cancel poller and write
directly.