pipe: add unread() operation
This patch adds an unread() operation to seastar::pipe, analogous to stdio's
ungetc(): After you unread() an item, the next read() will return it. The
typical use case is to read() an item, inspect it, and then unread() it
so it will be available again for the next read.

Note that unread() is a method of the pipe_reader object, not the pipe_writer
object (on which write() operates). In particular this means one cannot call
unread() in parallel with waiting on a read() - the pipe_reader object can only
be used by one fiber. This is fine, because the common use case will be, as
explained above, to call read() and only call unread() when the future returned
by read() is resolved.

In the current implementation, only a single item can be unread, which is enough
for the intended use case (and also, what ungetc() guarantees).

The implementation in this patch is nice and simple, but adds a small
performance penalty - one extra "if" on every read to check if there's any unread
item. This small penalty could have been avoided entirely by using an N+1-item
circular buffer, and unread() would push to its front. Such an implementation
is not difficult, but will require a much more extensive rewrite of pipe.hh,
and we can also do it later.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>