with_lock: run a function under a lock
This is meant, at least in principle, to be ran with the rwlock. We don't
really have many other types of lock, so perhaps this should be in the rwlock
file to begin with.

However, being a general facility in nature that allows us to "do something
with something else", let's put it in do_with.hh.

Signed-off-by: Glauber Costa <glommer@scylladb.com>