configure: don't fail when sanitizer is unavailable
We use try_compile_and_run to test the sanitizer's flags. But if the user
hasn't installed the sanitizer libraries at all, try_compile_and_run will
file with an exception...

The problem is that the code runs the compiler which asking it to write to
a temporary file, but the compiler fails and *deletes* that temporary file,
and then the temporary-file object's destructor throws an exception
(sadly, the Python guys haven't heard of the 11th commandment: thou shalt
not throw an exception from a destructor).

The fix is to create the temporary file again when the compilation failed.
Then the destructor will succeed in deleting it. Silly, but works.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>