input_stream: add read() method
To read from an input_stream, one normally uses the consume() method, which
is quite elaborate to set up. This patch introduces a simpler read() method,
returning a future temporary_buffer, containing some data or empty on an
end-of-stream.

A caller to read() has no control over how many characters will be read -
use read_exactly() if you need to read a specific number of characters.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>