output_stream: add write(temporary_buffer) variant
We already have a number of output_stream::write() overloads, but we missed
one taking a temporary_buffer. Having such a variant is convenient for
writing to some output_stream what you've just read from an input_stream.

The code is almost identical to the write(net::packet) variant which we
already had but was less useful because users do not normally have access
to net::packet.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>