configure.py: support for pkg-config before release 0.28
Older versions of pkg-config, before release 0.28 (see
http://lists.freedesktop.org/archives/pkg-config/2013-January/000912.html)
had a bug where duplicate compilation flags were deleted. This included
our use of -Wl,--whole-archive, which can apprear twice on our "Libs" list,
and one was removed. But one set of these flags surrounds the Seastar
libraries and one surrounds the DPDK libraries, and with these flags dropped
or reordered, compilation of a Seastar app using pkg-config will not work.

So this patch replaces the construct

    -Wl,--whole-archive -lseastar -Wl,--no-whole-archive

With the equivalent but space-free one:

    -Wl,--whole-archive,-lseastar,--no-whole-archive

pkg-config will now treat this whole thing as one single parameter, so it
will not remove any part of it as "duplicate" of anything else.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>