rpc: allow auxiliary data to be stored in client_info
Sometimes server's rpc handler wants to know some static information
about a client on behalf which it serves requests. There is not point in
passing such information with each request if it does not change during
client's lifetime. Allow a client to pass such information only once
and store it in client_info which is accessible in rpc handlers.