core: prevent explicit calls for to_sstring_sprintf()
The fact that to_sstring_sprintf() may be called
directly is dangerous even for integer values since a user
may provide a format like "%0500d" and cause to_sstring_sprintf()
internal buffer overrun.

This patch "hides" the to_sstring_sprintf() inside
the basic_sstring class by making it a static private method
so that it would be impossible to call it directly.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>