reactor: allow pollers to support an interrupt mode
Allow pollers to support interrupt mode by implementing a pair of methods:
try_enter_interrupt_mode() notifies the poller the shard is about to sleep
(and allows the poller to abort the entry into sleep by returning false).
exit_interrupt_mode() notifies the poller that the shard is switching back
into poll mode.

The poller will then take any actions it needs to switch the monitored
resource into a mode that allows it to wake the main loop.