reactor: interrupt mode support for smp queues
Introduce a reactor::_sleeping variable, and check it after sending a
message.  If the destination is asleep, wake it up with a signal.

Read-after-write tests like this require a full memory barrier, which
we don't insert here.  Instead, the code that sets _sleeping invokes
some magic to make sure all threads execute a barrier.  Since falling
asleep is rare compared to polling, the higher cost of the magic barrier
is compensated by the cheaper polling.