read/write_dma: do not pass "this" to lambda capture.
Within the IO queues framework, this function may not be executed in the
processor in which it was originated. To avoid complications, let us not pass
pointers that live in other CPUs.

We pass the "this" pointer so we can access the _fd member, and for nothing
else. We can easily avoid this by passing the fd alone in the capture.

Signed-off-by: Glauber Costa <glauber@scylladb.com>