reactor: define shard_id as a type
This is a simple typedef, since we don't want to prevent this to be used as
a standard unsigned - it is used as an index in many places to begin with.

The main goal is to allow us to make a clear statement about the role of
a particular unsigned when there is need to do it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>