tutorial.md: remove inaccurate statements about x86
Remove some inaccurate (and irrelevant) statements about PAE (which didn't
really offer 36 bits of virtual address space - just physical) and x86-64
(which perhaps theoretically supports 64 bit addresses, but actual
implementations are limited to 48 bits (and in Linux user-space only uses
half of those). But all of these details are really not relevant to
Seastar and we shouldn't waste precious time and space on them in the
tutorial.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>