reactor: increment check_direct_io_support to also deal with files
We use this check to make sure that a given directory supports Direct IO.
However, if we would like to use it to check if a file lives in a directory
that supports Direct IO, we would have to parse the file components and just
then call it with that directory as a parameter.

It is much better if we can try opening the file directly. Note that, while
calling open_file_dma() will already provide an exception if the file lives
in a directory that does not support Direct IO, the error message that we get
from the test is much more informative.

Signed-off-by: Glauber Costa <glauber@scylladb.com>