resource: commonize code for hwloc_distrib
The current call to hwloc_distrib is quite ugly: basically because the correct
API is version dependent, and we need to encode that into every caller. That is
fine when we have a single caller, but starts being a problem if we have more.

So let's encode it in a way that it can be easily reused later by other
callers.

To do that, I am creating a RAII object that can play the role of the previous
defer() method.

Signed-off-by: Glauber Costa <glauber@scylladb.com>