io queues: make the queues really part of the reactor
In Avi's latest review, he has asked me to move all_io_queues inside the
reactor as a static member. Which I though I did, except that I kept declaring
them in reactor.cc without the reactor:: specifier.

The effect of this is that there is now a static member of reactor that nobody
uses, and an outside member, which is the one actually used. This is not a bug
per se, but definetely not my initial intention.

Signed-off-by: Glauber Costa <glauber@scylladb.com>