reactor: use a vector of unique_ptr for the IO queues
The current design with an explicit new is an artifact of an earlier design
where the objects wouldn't be necessarily allocated in the correct processor.

Now they are, so we don't really need to use naked pointers, and anything like
that. Let's use unique_ptrs instead, without anything fancy around it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>