resources: make allocate return both the CPUs and IO queues
Currently we have two different functions to return both the IO queue and the CPU.
They share quite a bit of state, so we can gain a considerable amount by having a single
function allocating both.

Right now, all this patch does is some reorganization - moving allocate_io_queue upwards
so it can be static, and adjusting the return values.

In a later patch, we will integrate them tighter by sharing the topology.

Signed-off-by: Glauber Costa <glauber@scylladb.com>