submit_io: change pending flushes criteria
With the introduction of the IO queue mechanism, the current heuristic of
flushing async IO submissions to the kernel in batches of max_aio / 4 is less
than ideal.

For very slow disks, that can be significantly higher than what the IO queue
can hold.  This means that we will never submit the IO right away, but instead
always end up relying on the poller to do that for us.

Even more important than this, when the system is under pressure, meaning that
we have pending requests in the queue, we should start favoring latency over
batching and flush the requests right away to open up room for the pending ones
sooner rather than later.

Applying this patch during a ScyllaDB write workload in the presence of pending
IO dramatically reduces the amount of pending IO seen, as well as offers a slight
increase in disk throughput as a result.

Signed-off-by: Glauber Costa <glauber@scylladb.com>