do_with() optimization
do_with() is a commonly used idiom, so it is important that it is
optimal both in run-time and compile-time.

We used finally(), which is a construct over then_wrap() which adds
additional compile-time and currently, also runtime, overheads (see
issue #92 and issue #53).

do_with() uses finally() with a lambda returning a void (not a future)
and which cannot throw, and in this case much of the complexity of
finally() can be avoided, and it's trivial to switch to using then_wrapped
directly.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>