core: scollectd: Get rid of a clock_type alias
Get rid of a clock_type alias and introduce a 'duration'
alias (to std::chrono::milliseconds) instead.

scollectd code doesn't really care about the clock source
used by the timer as long as it's capable to consistently
generate periodical events (e.g. every 1s). And this is what
a default timer<> object is expected to provide.

It do cares however to use a duration type having the appropriate
resolution (not worse than 1ms). Therefore we introduce it explicitly
in this patch.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>