apps: memcached: use lowres_clock instead of a system_clock
The main reason for this patch is to avoid a system call for each added/updated
element with a non-zero expiration time in a form of a delta from the current time.
While on a physical machine time(0) is a memory read, in the environments
like AWS this would translate into a syscall. Therefore there is a good
motivation to avoid this.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>