apps: memcached: Don't fiddle with the expiration time value in a usable range
This patch fixes a regression introduced by a
06126cf14cbb0eb8446ca8fb45dd9d25d950511b commit.

"expiration" class is used not only for handling the elements' expiration
field but also for triggering "flush" commands. Therefore we may not
change the requested time point when it has a value which may actually
be programmed by a uint32_t value according to a memcached protocol, e.g.
a time point in the past that may be programmed by an absolute time notation.

However setting a "never expire" value (based on an int64_t) to be big/small
enough so that it could not collide with the uint32_t based expiration values
would do the job.

This patch reverts 06126cf14cbb0eb8446ca8fb45dd9d25d950511b and sets
a "never expire" value to LLONG_MIN. This ensures that a minimal value in
_time will not collide with the "never expire" value for about next 290k years.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>