fix operator<<() for std::vector<T>
core/sstring.hh has an operator<<() implementation for std::vector<T>.

It is debatable why Seastar should have a pretty-printer only for this
specific container and not others - and if it does, why it is in sstring.hh
(I started a thread on the mailing list to discuss this). But as long
as we do have this operator, let's do it correctly.

One of the many complications of C++ is that an operator<<() overload
needs to be in the same namespace as its second parameter. Specifically,
if we're trying here to define operator<< for an std::vector, this
definition must be in the "std" namespace. The problems caused by
forgetting this are subtle: "<<" will work from main-namespace code,
but will *not* work from code in any other namespace. In particular,
any code which prints through sprint() or Boost's printing code (in
the Boost namespace), will *not* be able to print vectors, because this
overload will not be noticed!

This patch adds the missing "namespace std", and fixes this problem:
vectors can now be used in sprint().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>