tests: memcached: workaround a possible race between flush_all and read
Currently memcached will return "OK" for flush_all after it
scheduled a timer that is going to evict all elements from the cache.
However if this operation came in the middle of the timer period
(e.g. timer's granularity is 1ms and the operation came 50us after the last
timer event) it allows read operation to arrive and to be handled before
the next timer event and thereby cause a flush_in_the_past test to fail.

There is an issue with our flush_all implementation in general
(see https://github.com/scylladb/seastar/issues/101 for more details).

This workaround would allow the test to pass before we get to fixing
the issue above.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Message-Id: <1452012791-4410-3-git-send-email-vladz@cloudius-systems.com>