IO queues: destruct within reactor's destructor
Our IO queues currently live as unique_ptrs in a static vector inside the
reactor. They are shard-local in creation, but if we don't destroy them
ourselves, they may be destroyed in a different shard when the reactor is
destroyed - as static members go separately for that.

What we have here works fine, but as soon as we start stashing more complex
objects into the IO queue, like shared pointers that will explicitly check
sharding when decrementing refcounts, we start to get into trouble.

In preparation to that, we will change the way we store IO queues slightly.
Each queue will be stored as an unique_ptr inside the reactor that belongs
to the shard responsible for creating it, which guarantees it will be destructed
at the right context.

The all_io_queues vector can now be a local vector containing just the raw
pointers, that lives just for as long as the configure function lives. A bit
more care must be taken to make sure it will not be accessed by the other
threads after the function has exited. But we have more then enough
synchronization points in place to guarantee it: we just need to reorder
the calls a bit.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1452105910-7941-1-git-send-email-glauber@scylladb.com>