fstream: add a priority class parameter
We have a priority class parameter available for the read / write dma
interface.  However, much of our IO happens through the fstream interface
instead.

For the benefit of existing users (that may live out of tree), we will keep
the interface that does not pass any priority class available - and that will
use the default priority class.

Signed-off-by: Glauber Costa <glauber@scylladb.com>