io_queue: remove pending_io counter
The usefulness of this counter was overstated by me when I first proposed its
inclusion. In practice, I haven't used it even once when looking for problems -
usually only looking at the number of pending operations tend to be enough.

It wouldn't be such a problem if this was zero impact, but the way it is, it
forces us to issue an extra continuation per each request.

I propose we remove it, and add it back should we really be able to justify its
need in the future.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <1453144380-30053-1-git-send-email-glauber@scylladb.com>